workers/requirements.txt
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.14.0
flask==3.0.0
gunicorn==21.2.0
